requirements.txt
fastapi
uvicorn
ytmusicapi
jinja2
python-multipart
spotipy
python-dotenv
